Re-structuring

+++ test/test_vectors.hpp
@@ -1,11 +1,11 @@
 #ifndef _TEST_VECTORS_HPP_
 #define _TEST_VECTORS_HPP_
 
+#include <cstdint>
+
 #include <vector>
 
-#include <stdint.h>
-
 extern std::vector<uint8_t *> vectors;
-extern std::vector<int> vectors_sizes;
+extern std::vector<size_t> vectors_sizes;
 
 #endif  // _TEST_VECTORS_HPP_
